# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SplitWiseMoney API",
                 "description":  "SplitWiseMoney - Expense sharing web application API documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/expenses/{id}":  {
                                             "put":  {
                                                         "tags":  [
                                                                      "Expenses"
                                                                  ],
                                                         "summary":  "Edit an existing expense",
                                                         "operationId":  "editExpense",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/ExpenseRequest"
                                                                                                                                 }
                                                                                                                  }
                                                                                         },
                                                                             "required":  true
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "*/*":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/ExpenseResponse"
                                                                                                                              }
                                                                                                               }
                                                                                                   }
                                                                                   }
                                                                       },
                                                         "security":  [
                                                                          {
                                                                              "bearerAuth":  [

                                                                                             ]
                                                                          }
                                                                      ]
                                                     },
                                             "delete":  {
                                                            "tags":  [
                                                                         "Expenses"
                                                                     ],
                                                            "summary":  "Delete an expense",
                                                            "operationId":  "deleteExpense",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int64"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK",
                                                                                          "content":  {
                                                                                                          "*/*":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "string"
                                                                                                                                 }
                                                                                                                  }
                                                                                                      }
                                                                                      }
                                                                          },
                                                            "security":  [
                                                                             {
                                                                                 "bearerAuth":  [

                                                                                                ]
                                                                             }
                                                                         ]
                                                        }
                                         },
...
                  "/api/expenses/group/{groupId}":  {
                                                        "get":  {
                                                                    "tags":  [
                                                                                 "Expenses"
                                                                             ],
                                                                    "summary":  "Get paginated list of group expenses",
                                                                    "operationId":  "getGroupExpenses",
                                                                    "parameters":  [
                                                                                       {
                                                                                           "name":  "groupId",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "integer",
                                                                                                          "format":  "int64"
                                                                                                      }
                                                                                       },
                                                                                       {
                                                                                           "name":  "page",
                                                                                           "in":  "query",
                                                                                           "required":  false,
                                                                                           "schema":  {
                                                                                                          "type":  "integer",
                                                                                                          "format":  "int32",
                                                                                                          "default":  0
                                                                                                      }
                                                                                       },
                                                                                       {
                                                                                           "name":  "size",
                                                                                           "in":  "query",
                                                                                           "required":  false,
                                                                                           "schema":  {
                                                                                                          "type":  "integer",
                                                                                                          "format":  "int32",
                                                                                                          "default":  10
                                                                                                      }
                                                                                       },
                                                                                       {
                                                                                           "name":  "sort",
                                                                                           "in":  "query",
                                                                                           "required":  false,
                                                                                           "schema":  {
                                                                                                          "type":  "string",
                                                                                                          "default":  "expenseDate,desc"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK",
                                                                                                  "content":  {
                                                                                                                  "*/*":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/PageExpenseResponse"
                                                                                                                                         }
                                                                                                                          }
                                                                                                              }
                                                                                              }
                                                                                  },
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ]
                                                                },
                                                        "post":  {
                                                                     "tags":  [
                                                                                  "Expenses"
                                                                              ],
                                                                     "summary":  "Add a new expense to a group",
                                                                     "operationId":  "addExpense",
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "groupId",
                                                                                            "in":  "path",
                                                                                            "required":  true,
                                                                                            "schema":  {
                                                                                                           "type":  "integer",
                                                                                                           "format":  "int64"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "requestBody":  {
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/ExpenseRequest"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     },
                                                                                         "required":  true
                                                                                     },
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "OK",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/ExpenseResponse"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               }
                                                                                   },
                                                                     "security":  [
                                                                                      {
                                                                                          "bearerAuth":  [

                                                                                                         ]
                                                                                      }
                                                                                  ]
                                                                 }
                                                    },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "ExpenseRequest":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "amount":  {
                                                                                                "type":  "number",
                                                                                                "minimum":  0.01
                                                                                            },
                                                                                 "description":  {
                                                                                                     "type":  "string",
                                                                                                     "maxLength":  255,
                                                                                                     "minLength":  0
                                                                                                 },
                                                                                 "category":  {
                                                                                                  "type":  "string",
                                                                                                  "maxLength":  50,
                                                                                                  "minLength":  0
                                                                                              },
                                                                                 "expenseDate":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "date"
                                                                                                 },
                                                                                 "paidById":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int64"
                                                                                              },
                                                                                 "participantShares":  {
                                                                                                           "type":  "object",
                                                                                                           "additionalProperties":  {
                                                                                                                                        "type":  "number"
                                                                                                                                    }
                                                                                                       }
                                                                             },
                                                              "required":  [
                                                                               "amount",
                                                                               "category",
                                                                               "description",
                                                                               "expenseDate",
                                                                               "paidById",
                                                                               "participantShares"
                                                                           ]
                                                          },
                                       "ExpenseResponse":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int64"
                                                                                         },
                                                                                  "amount":  {
                                                                                                 "type":  "number"
                                                                                             },
                                                                                  "description":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                  "category":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                  "expenseDate":  {
                                                                                                      "type":  "string",
                                                                                                      "format":  "date"
                                                                                                  },
                                                                                  "paidById":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               },
                                                                                  "paidByName":  {
                                                                                                     "type":  "string"
                                                                                                 },
                                                                                  "participants":  {
                                                                                                       "type":  "array",
                                                                                                       "items":  {
                                                                                                                     "$ref":  "#/components/schemas/ParticipantShareDto"
                                                                                                                 }
                                                                                                   }
                                                                              }
                                                           },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}